feat(backend): Add automapper to backend project
Add automapper to backend project, with an example mapping profile to map Review
entities to DTOs (ReviewMappingProfle.cs). We can inject an IMapper instance
into our services, and Automapper with apply the profile automatically. We
will then be able to call IMapper.Map to map our entities automatically based on
that profile.