feat: Modify components to use new Http Services
This was a simple switch to using injected services, rather than
creating a new Backend.cs instance in each. The only significant change
was that the services return ServiceObjectResponse, and the value must
be read from these. In the future, it would be good use the
ServiceObjectResponse Type and Messages properties to handle errors.