Update comment functionality
Updated comment adding functionality so that comments are added and displayed in real time. Add comment view and edit buttons so that the form and comments can be selectively hidden on demand. Also changed the display conditionals so that the correct message displays when there are no reviews on a specific episode.