Add edit review functionality and ensure real-time updating
Added functionality to edit reviews using an edit review popup and connecting to the put request on the backend. Had to add a refreshreviews cascading parameter, which is a class with a refresh bool and onrefreshchanged action to implement StatehasChanged in the main layout. This ensures that the cascading value is updated and the reviews are reloaded. Applied this to both add review and edit review. Will do the same for deleting reviews.